+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_template_builder.py
@@ -36,7 +36,7 @@ def build_application_gateway_resource(cmd, name, location, tags, sku_name, sku_
                                        private_ip_address, private_ip_allocation, cert_data, cert_password,
                                        cookie_based_affinity, http_settings_protocol, http_settings_port,
                                        http_listener_protocol, routing_rule_type, public_ip_id, subnet_id,
-                                       connection_draining_timeout, enable_http2):
+                                       connection_draining_timeout, enable_http2, min_capacity):
 
     # set the default names
     frontend_ip_name = 'appGatewayFrontendIP'
@@ -142,6 +142,9 @@ def build_application_gateway_resource(cmd, name, location, tags, sku_name, sku_
         ag_properties.update({'sslCertificates': [ssl_cert]})
     if enable_http2 and cmd.supported_api_version(min_api='2017-10-01'):
         ag_properties.update({'enableHttp2': enable_http2})
+    if cmd.supported_api_version(min_api='2018-07-01'):
+        ag_properties.update({'autoscaleConfiguration': {'minCapacity': min_capacity}})
+        del ag_properties['sku']['capacity']
 
     ag = {
         'type': 'Microsoft.Network/applicationGateways',