+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/custom.py
@@ -156,7 +156,7 @@ def create_application_gateway(cmd, application_gateway_name, resource_group_nam
                                subnet='default', subnet_address_prefix='10.0.0.0/24',
                                virtual_network_name=None, vnet_address_prefix='10.0.0.0/16',
                                public_ip_address_type=None, subnet_type=None, validate=False,
-                               connection_draining_timeout=0, enable_http2=None):
+                               connection_draining_timeout=0, enable_http2=None, min_capacity=2):
     from azure.cli.core.util import random_string
     from azure.cli.core.commands.arm import ArmTemplateBuilder
     from azure.cli.command_modules.network._template_builder import (
@@ -207,7 +207,7 @@ def create_application_gateway(cmd, application_gateway_name, resource_group_nam
         private_ip_address, private_ip_allocation, cert_data, cert_password,
         http_settings_cookie_based_affinity, http_settings_protocol, http_settings_port,
         http_listener_protocol, routing_rule_type, public_ip_id, subnet_id,
-        connection_draining_timeout, enable_http2)
+        connection_draining_timeout, enable_http2, min_capacity)
     app_gateway_resource['dependsOn'] = ag_dependencies
     master_template.add_variable(
         'appGwID',
@@ -232,7 +232,7 @@ def create_application_gateway(cmd, application_gateway_name, resource_group_nam
     return sdk_no_wait(no_wait, client.create_or_update, resource_group_name, deployment_name, properties)
 
 
-def update_application_gateway(instance, sku=None, capacity=None, tags=None, enable_http2=None):
+def update_application_gateway(instance, sku=None, capacity=None, tags=None, enable_http2=None, min_capacity=None):
     if sku is not None:
         instance.sku.name = sku
         instance.sku.tier = sku.split('_', 1)[0] if 'v2' not in sku else sku
@@ -242,6 +242,8 @@ def update_application_gateway(instance, sku=None, capacity=None, tags=None, ena
         instance.tags = tags
     if enable_http2 is not None:
         instance.enable_http2 = enable_http2
+    if min_capacity is not None:
+        instance.autoscale_configuration.min_capacity = min_capacity
     return instance
 
 
@@ -635,6 +637,26 @@ def set_ag_ssl_policy_2017_06_01(cmd, resource_group_name, application_gateway_n
 def show_ag_ssl_policy(cmd, resource_group_name, application_gateway_name):
     return network_client_factory(cmd.cli_ctx).application_gateways.get(
         resource_group_name, application_gateway_name).ssl_policy
+
+
+def create_ag_trusted_root_certificate(cmd, resource_group_name, application_gateway_name, item_name, no_wait=False,
+                                       cert_data=None, keyvault_secret=None):
+    ApplicationGatewayTrustedRootCertificate = cmd.get_models('ApplicationGatewayTrustedRootCertificate')
+    ncf = network_client_factory(cmd.cli_ctx).application_gateways
+    ag = ncf.get(resource_group_name, application_gateway_name)
+    root_cert = ApplicationGatewayTrustedRootCertificate(name=item_name, data=cert_data,
+                                                         keyvault_secret_id=keyvault_secret)
+    _upsert(ag, 'trusted_root_certificates', root_cert, 'name')
+    return sdk_no_wait(no_wait, ncf.create_or_update,
+                       resource_group_name, application_gateway_name, ag)
+
+
+def update_ag_trusted_root_certificate(instance, parent, item_name, cert_data=None, keyvault_secret=None):
+    if cert_data is not None:
+        instance.data = cert_data
+    if keyvault_secret is not None:
+        instance.keyvault_secret_id = keyvault_secret
+    return parent
 
 
 def create_ag_url_path_map(cmd, resource_group_name, application_gateway_name, item_name, paths,